Initial support for 64-bit file lengths, untested. Likely issues with use of printf() and sscanf().

+++ src/sbbs3/zmodem.h
@@ -229,15 +229,15 @@ typedef struct {
 	BYTE rx_data_subpacket[8192];							/* zzap = 8192 */
 
 	char		current_file_name[MAX_PATH+1];
-	uint32_t	current_file_size;
-	uint32_t	current_file_pos;
+	int64_t		current_file_size;
+	int64_t		current_file_pos;
 	time_t		current_file_time;
 	unsigned	current_file_num;
 	unsigned	total_files;
-	uint32_t	total_bytes;
+	int64_t		total_bytes;
 	unsigned	files_remaining;
-	uint32_t	bytes_remaining;
-	uint32_t	transfer_start_pos;
+	int64_t		bytes_remaining;
+	int64_t		transfer_start_pos;
 	time_t		transfer_start_time;
 
 	int		receive_32bit_data;
@@ -271,15 +271,16 @@ typedef struct {
 	unsigned	max_errors;
 	unsigned	block_size;
 	unsigned	max_block_size;
-	uint32_t	max_file_size;		/* 0 = unlimited */
+	int64_t		max_file_size;		/* 0 = unlimited */
 	int			*log_level;
 
 	/* Callbacks */
+	/* error C2520: conversion from unsigned __int64 to double not implemented, use signed __int64 */
 	void*		cbdata;
 	int			(*lputs)(void*, int level, const char* str);
 	int			(*send_byte)(void*, BYTE ch, unsigned timeout);
 	int			(*recv_byte)(void*, unsigned timeout);
-	void		(*progress)(void*, uint32_t current_pos);
+	void		(*progress)(void*, int64_t current_pos);
 	BOOL		(*is_connected)(void*);
 	BOOL		(*is_cancelled)(void*);
 	BOOL		(*data_waiting)(void*, unsigned timeout);
@@ -290,7 +291,7 @@ typedef struct {
 
 void		zmodem_init(zmodem_t*, void* cbdata
 						,int	(*lputs)(void*, int level, const char* str)
-						,void	(*progress)(void*, uint32_t current_pos)
+						,void	(*progress)(void*, int64_t current_pos)
 						,int	(*send_byte)(void*, BYTE ch, unsigned timeout)
 						,int	(*recv_byte)(void*, unsigned timeout)
 						,BOOL	(*is_connected)(void*)
@@ -314,10 +315,10 @@ BOOL		zmodem_get_crc(zmodem_t*, int32_t length, uint32_t* crc);
 void		zmodem_parse_zrinit(zmodem_t*);
 void		zmodem_parse_zfile_subpacket(zmodem_t* zm);
 int			zmodem_send_zfin(zmodem_t*);
-BOOL		zmodem_send_file(zmodem_t*, char* name, FILE* fp, BOOL request_init, time_t* start, uint32_t* bytes_sent);
-int			zmodem_recv_files(zmodem_t* zm, const char* download_dir, uint32_t* bytes_received);
+BOOL		zmodem_send_file(zmodem_t*, char* name, FILE* fp, BOOL request_init, time_t* start, int64_t* bytes_sent);
+int			zmodem_recv_files(zmodem_t* zm, const char* download_dir, int64_t* bytes_received);
 int			zmodem_recv_init(zmodem_t* zm);
-unsigned	zmodem_recv_file_data(zmodem_t*, FILE*, uint32_t offset);
+unsigned	zmodem_recv_file_data(zmodem_t*, FILE*, int64_t offset);
 int			zmodem_recv_file_frame(zmodem_t* zm, FILE* fp);
 int			zmodem_recv_header_and_check(zmodem_t* zm);
 #endif
